backend/requirements.txt
  Django
  djangorestframework
  django-cors-headers
  scikit-learn
  pandas